requirements.txt
langchain
groq
google-generativeai
python-dotenv
streamlit
#./en_core_sci_sm-0.5.4.tar.gz
#https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz
#python -m spacy download en_core_web_sm
xhtml2pdf
fastapi
uvicorn[standard]
reportlab


